+++ palo/panorama/panorama_export.py
@@ -55,7 +55,7 @@ Usage:
 License: MIT
 """
 
-__version__ = "1.6.1"
+__version__ = "1.7.0"
 
 import argparse
 import re